continue refactoring module structure
