+++ src/cfn/formatting.ts
@@ -1,12 +1,14 @@
-import * as _ from 'lodash';
-import * as aws from 'aws-sdk'
+import * as aws from 'aws-sdk';
 import * as cli from 'cli-color';
 import * as dateformat from 'dateformat';
+import * as _ from 'lodash';
 import {sprintf} from 'sprintf-js';
+import {FAILURE, SUCCESS} from '../statusCodes';
 
 export const COLUMN2_START = 25;
 export const DEFAULT_STATUS_PADDING = 35;
 export const MIN_STATUS_PADDING = 17;
+export const MAX_PADDING = 60;
 
 export function renderTimestamp(ts: Date) {
   return dateformat(ts);
@@ -24,6 +26,10 @@ export const printSectionEntry = (label: string, data: string): boolean =>
 export const formatLogicalId = (s: string) => cli.xterm(252)(s);
 export const formatStackOutputName = formatLogicalId;
 export const formatStackExportName = formatLogicalId;
+
+export function calcPadding<T>(items: T[], selector: (x:T) => string): number {
+  return Math.min(_.max(_.map(items, i => selector(i).length)) as number, MAX_PADDING);
+}
 
 export function colorizeResourceStatus(status: string, padding = DEFAULT_STATUS_PADDING): string {
   padding = (_.isNumber(padding) && padding >= MIN_STATUS_PADDING) ? padding : MIN_STATUS_PADDING;
@@ -89,4 +95,17 @@ export const prettyFormatTags = (tags?: aws.CloudFormation.Tags): string => {
     return '';
   }
   return prettyFormatSmallMap(_.fromPairs(_.map(tags, (tag) => [tag.Key, tag.Value])));
+}
+
+
+export function showFinalComandSummary(wasSuccessful: boolean): number {
+  if (wasSuccessful) {
+    console.log(formatSectionHeading(sprintf(`%-${COLUMN2_START}s`, 'Command Summary:')),
+                cli.black(cli.bgGreenBright('Success')), '👍');
+    return SUCCESS;
+  } else {
+    console.log(formatSectionHeading(sprintf(`%-${COLUMN2_START}s`, 'Command Summary:')),
+                cli.bgRedBright('Failure'), ' (╯°□°）╯︵ ┻━┻ ', 'Fix and try again.');
+    return FAILURE;
+  }
 }
