+++ src/cfn/index.ts
@@ -1,671 +1,30 @@
 import * as aws from 'aws-sdk';
 import * as cli from 'cli-color';
-import * as fs from 'fs';
 import * as _ from 'lodash';
-import * as pathmod from 'path';
-import * as process from 'process';
-import didYouMean from 'didyoumean2';
 import * as nameGenerator from 'project-name-generator';
 import * as querystring from 'querystring';
-import * as request from 'request-promise-native';
-import {sprintf} from 'sprintf-js';
-
-import {GenericCLIArguments} from '../cli';
-import * as yaml from '../yaml';
-import {logger} from '../logger';
-import getCurrentAWSRegion from '../getCurrentAWSRegion';
-import configureAWS from '../configureAWS';
-import {Md5} from 'ts-md5/dist/md5';
-import * as url from 'url';
-import {AWSRegion} from '../aws-regions';
 import calcElapsedSeconds from '../calcElapsedSeconds';
+import {GenericCLIArguments} from '../cli-util';
 import confirmationPrompt from '../confirmationPrompt';
-import def from '../default';
-import {ExtendedCfnDoc, PreprocessOptions, readFromImportLocation, transform} from '../preprocess';
+import getCurrentAWSRegion from '../getCurrentAWSRegion';
+import {logger} from '../logger';
 import mkSpinner from '../spinner';
 import {FAILURE, INTERRUPT, SUCCESS} from '../statusCodes';
 import timeout from '../timeout';
+import {AbstractCloudFormationStackCommand} from './AbstractCloudFormationStackCommand';
 import {diffStackTemplates} from './diffStackTemplates';
-import {
-  COLUMN2_START,
-  formatSectionHeading,
-  formatTimestamp,
-  prettyFormatSmallMap,
-  printSectionEntry,
-  renderTimestamp
-} from './formatting';
-import {getAllStackEvents} from './getAllStackEvents';
-import getReliableStartTime from './getReliableStartTime';
-import maybeSignS3HttpUrl from './maybeSignS3HttpUrl';
-import objectToCFNParams from './objectToCFNParams';
-import objectToCFNTags from './objectToCFNTags';
-import {runCommandSet} from './runCommandSet';
-import {showStackEvents} from './showStackEvents';
+import {formatSectionHeading, showFinalComandSummary} from './formatting';
+import {loadStackArgs} from './loadStackArgs';
+import {stackArgsToCreateChangeSetInput} from './stackArgsToX';
 import {summarizeStackDefinition} from './summarizeStackDefinition';
-import {getStackDescription} from './getStackDescription';
-import {watchStack} from './watchStack';
-import {StackArgs, CfnOperation} from './types';
-import {extendedCfnDocKeys} from '../preprocess/visitor';
-import {filter} from '../preprocess/filter';
-import {summarizeStackContents} from './summarizeStackContents';
 import terminalStackStates from './terminalStackStates';
-import {DEFAULT_EVENT_POLL_INTERVAL} from './defaults';
-
-const stackArgsProperties: Array<keyof StackArgs> = [
-  'ApprovedTemplateLocation',
-  'AssumeRoleARN',
-  'Capabilities',
-  'ClientRequestToken',
-  'CommandsBefore',
-  'DisableRollback',
-  'EnableTerminationProtection',
-  'NotificationARNs',
-  'OnFailure',
-  'Parameters',
-  'Profile',
-  'Region',
-  'ResourceTypes',
-  'RoleARN',
-  'ServiceRoleARN',
-  'StackName',
-  'StackPolicy',
-  'Tags',
-  'Template',
-  'TimeoutInMinutes',
-  'UsePreviousTemplate',
-];
-
-export async function showPendingChangesets(StackId: string, changeSetsPromise?: Promise<aws.CloudFormation.ListChangeSetsOutput>) {
-  const cfn = new aws.CloudFormation();
-  if (!changeSetsPromise) {
-    changeSetsPromise = cfn.listChangeSets({StackName: StackId}).promise();
-  }
-  // TODO pagination if lots of changesets:
-  let changeSets = def([], (await changeSetsPromise).Summaries);
-  changeSets = _.sortBy(changeSets, (cs) => cs.CreationTime);
-  if (changeSets.length > 0) {
-    console.log();
-    console.log(formatSectionHeading('Pending Changesets:'))
-    for (const cs of changeSets) {
-      printSectionEntry(
-        formatTimestamp(renderTimestamp(cs.CreationTime as Date)),
-        cli.magenta(cs.ChangeSetName) +
-        ' ' +
-        cs.Status +
-        ' ' +
-        def('', cs.StatusReason));
-      if (!_.isEmpty(cs.Description)) {
-        console.log('  Description:', cli.blackBright(cs.Description))
-        console.log()
-      }
-      summarizeChangeSet(await cfn.describeChangeSet({StackName: StackId, ChangeSetName: cs.ChangeSetName!}).promise());
-      console.log()
-    }
-  }
-}
-
-async function loadCFNStackPolicy(policy: string | object | undefined, baseLocation: string):
-  Promise<{StackPolicyBody?: string, StackPolicyURL?: string}> {
-
-  if (_.isUndefined(policy)) {
-    return {};
-  } else if (_.isString(policy)) {
-    const location0 = policy;
-    const shouldRender = (location0.trim().indexOf('render:') === 0);
-    const location = maybeSignS3HttpUrl(location0.trim().replace(/^ *render: */, ''));
-    const importData = await readFromImportLocation(location, baseLocation);
-    if (!shouldRender && importData.importType === 's3') {
-      throw new Error(`Use https:// urls when using a plain (non-rendered) StackPolicy from S3: ${location}`);
-      // note, s3 urls are valid for the shouldRender case below
-    } else if (!shouldRender && importData.importType === 'http') {
-      return {StackPolicyURL: importData.resolvedLocation};
-    } else {
-      return {
-        StackPolicyBody: shouldRender
-          ? JSON.stringify(await transform(importData.doc, importData.resolvedLocation), null, ' ')
-          : importData.data
-      };
-    }
-  } else if (_.isObject(policy)) {
-    return {StackPolicyBody: JSON.stringify(policy)}
-  } else {
-    return {}
-  }
-}
-
-const TEMPLATE_MAX_BYTES = 51199
-export async function loadCFNTemplate(location0: string, baseLocation: string, environment: string, options: PreprocessOptions = {}):
-  Promise<{TemplateBody?: string, TemplateURL?: string}> {
-  if (_.isUndefined(location0)) {
-    return {};
-  }
-  const shouldRender = (location0.trim().indexOf('render:') === 0);
-  const location = maybeSignS3HttpUrl(location0.trim().replace(/^ *render: */, ''));
-  // We auto-sign any s3 http urls here ^ prior to reading from them
-  // (via readFromImportLocation below) or passing them to CFN via
-  // TemplateUrl. This allows iidy to handle cross-region
-  // TemplateUrls. s3:// urls don't provide any means of encoding
-  // the source region and CFN doesn't accept them.
-
-  // TODO maybeSignS3HttpUrl might need updating later if we add
-  // support for baseLocation here being an http url itself: i.e.
-  // relative imports. This is probably an edge-case we don't need to
-  // support but it's worth noting.
-
-  if (!shouldRender && location.match(/^s3:/)) {
-    throw new Error(`Use https:// urls when using a plain (non-rendered) Template from S3: ${location}`);
-    // note, s3 urls are valid for the shouldRender case below
-  } else if (!shouldRender && location.match(/^http/)) {
-    // note the handling of unsigned s3 http urls above in maybeSignS3HttpUrl ^
-    return {TemplateURL: location};
-  } else {
-    const importData = await readFromImportLocation(location, baseLocation);
-    if (importData.data.indexOf('$imports:') > -1 && !shouldRender) {
-      throw new Error(
-        `Your cloudformation Template from ${location} appears to`
-        + ' use iidy\'s yaml pre-processor syntax.\n'
-        + ' You need to prefix the template location with "render:".\n'
-        + ` e.g.   Template: "render:${location}"`
-      );
-    }
-
-    importData.doc.$envValues = _.merge({}, importData.doc.$envValues, {iidy: {environment: environment, region: getCurrentAWSRegion()}});
-    const body = shouldRender
-      ? yaml.dump(await transform(importData.doc, importData.resolvedLocation, options))
-      : importData.data;
-    if (body.length >= TEMPLATE_MAX_BYTES) {
-      throw new Error('Your cloudformation template is larger than the max allowed size. '
-        + 'You need to upload it to S3 and reference it from there.')
-    }
-    return {TemplateBody: body};
-  }
-}
-
-export async function addDefaultNotificationArn(args: StackArgs): Promise<StackArgs> {
-  const ssm = new aws.SSM();
-  const ssmLookup = await ssm.getParameter(
-    {Name: '/iidy/default-notification-arn', WithDecryption: true}).promise().catch(() => null);
-  if (ssmLookup && ssmLookup.Parameter && ssmLookup.Parameter.Value) {
-    const TopicArn = ssmLookup.Parameter.Value;
-    const sns = new aws.SNS();
-    if (await sns.getTopicAttributes({TopicArn}).promise().return(true).catchReturn(false)) {
-      args.NotificationARNs = (args.NotificationARNs || []).concat(TopicArn);
-    } else {
-      logger.warn(
-        `iidy's default NotificationARN set in this region is invalid: ${TopicArn}`);
-    }
-  }
-  return args;
-}
-
-export async function loadStackArgs(argv: GenericCLIArguments,
-                                    filterKeys: string[] = [],
-                                    setupAWSCredentails = configureAWS): Promise<StackArgs> {
-  // TODO json schema validation
-  const args = await _loadStackArgs(argv.argsfile, argv, filterKeys, setupAWSCredentails);
-  if (argv.clientRequestToken) {
-    args.ClientRequestToken = argv.clientRequestToken;
-  }
-  return addDefaultNotificationArn(args);
-}
-
-function showArgsfileWarnings(argsdata: object, filename: string) {
-  const invalidProperties = _.difference(_.keys(argsdata),
-                                         stackArgsProperties,
-                                         extendedCfnDocKeys);
-  _.forEach(invalidProperties, (name: string) => {
-    let suggestion = '';
-    const suggestedProperty = didYouMean(name, stackArgsProperties);
-    if (suggestedProperty) {
-      suggestion = `. Did you mean '${suggestedProperty}'?`
-    }
-    logger.warn(`Invalid property '${name}' in ${filename}${suggestion}`);
-  });
-}
-
-//export async function _loadStackArgs(argsfile: string, region?: AWSRegion, profile?: string, environment?: string): Promise<StackArgs> {
-export async function _loadStackArgs(argsfile: string,
-                                     argv: GenericCLIArguments,
-                                     filterKeys: string[] = [],
-                                     setupAWSCredentails = configureAWS): Promise<StackArgs> {
-  const profile: string | undefined = argv.profile;
-  const assumeRoleArn: string | undefined = argv.assumeRoleArn;
-  const environment: string | undefined = argv.environment;
-  const iidy_command = argv._.join(' ');
-
-  let argsdata: any; // tslint:disable-line
-  if (!fs.existsSync(argsfile)) {
-    throw new Error(`stack args file "${argsfile}" not found`);
-  } else if (pathmod.extname(argsfile) === '.json') {
-    argsdata = JSON.parse(fs.readFileSync(argsfile).toString());
-  } else if (_.includes(['.yaml', '.yml'], pathmod.extname(argsfile))) {
-    argsdata = yaml.loadString(fs.readFileSync(argsfile), argsfile)
-  } else {
-    throw new Error(`Invalid stack args file "${argsfile}" extension`);
-  }
-
-  if(!_.isEmpty(filterKeys)) {
-    argsdata = filter(filterKeys, argsdata, argsfile);
-  }
-
-  // There is chicken-and-egg situation between use of imports for
-  // profile or region and the call to configureAWS. We need to
-  // enforce that they be plain strings with no pre-processor magic.
-  for (const key of ['Profile', 'AssumeRoleARN', 'Region']) {
-    if (_.isObject(argsdata[key])) {
-      if (environment && argsdata[key][environment]) {
-        argsdata[key] = argsdata[key][environment];
-        logger.debug(`resolving ${key}=${argsdata[key]} based on environment=${environment}`);
-      } else {
-        throw new Error(`environment "${environment}" not found in ${key} map: ${argsdata[key]}`);
-      }
-    } else if (argsdata[key] && !_.isString(argsdata[key])) {
-      throw new Error(`The ${key} setting in stack-args.yaml must be a plain string or an environment map of strings.`);
-    }
-  }
-  // have to configureAws before the call to transform as $imports might make AWS api calls.
-  const cliOptionOverrides = _.pickBy(argv, (v: any, k: string) => !_.isEmpty(v) && _.includes(['region', 'profile', 'assumeRoleArn'], k));
-  const argsfileSettings = {profile: argsdata.Profile, assumeRoleArn: argsdata.AssumeRoleARN, region: argsdata.Region};
-  const mergedAWSSettings = _.merge(argsfileSettings, cliOptionOverrides);
-
-  logger.debug(`loadStackArgs cliOptionOverrides`, cliOptionOverrides);
-  logger.debug(`loadStackArgs argsfileSettings`, argsfileSettings);
-  logger.debug(`loadStackArgs mergedAWSSettings`, mergedAWSSettings);
-
-  await setupAWSCredentails(mergedAWSSettings); // cliOptionOverrides trump argsfile
-
-  if (environment) {
-    if (!_.get(argsdata, ['Tags', 'environment'])) {
-      argsdata.Tags = _.merge({environment}, argsdata.Tags);
-    }
-  }
-  const finalRegion = getCurrentAWSRegion();
-  argsdata.$envValues = _.merge(
-    {}, argsdata.$envValues, {
-      // TODO deprecate bare region/environment:
-      region: finalRegion,
-      environment,
-      // new style with namespace to avoid clashes:
-      iidy: {
-        command: iidy_command,
-        environment,
-        region: finalRegion,
-        profile: mergedAWSSettings.profile
-      }
-    });
-
-  if (argsdata.CommandsBefore) {
-    // TODO should we actually execute the commands if this is `iidy render`?
-    if (_.includes(['create-stack', 'update-stack', 'create-changeset', 'create-or-update'], iidy_command)) {
-      // The CommandsBefore strings are pre-processed for any handlebars
-      // templates they contain. We call `transform` once here to get
-      // the $envValues ($imports, $defs, etc.) and fully rendered
-      // StackArgs so they're available to handlebars. It's called again
-      // below to produce the final `stackArgsPass2` as these commands
-      // might alter the values in $imports. For example, an import of
-      // `filehash:lambda.zip` would change after the
-      //
-      const argsdataPass1: ExtendedCfnDoc = _.omit(_.cloneDeep(argsdata), ['CommandsBefore']);
-      // NOTE any AWS api calls made in the imports will be made twice
-      // because of the multiple passes. TODO use transformPostImports
-      // instead and loadImports only once.
-      const stackArgsPass1 = await transform(argsdataPass1, argsfile) as StackArgs;
-
-      // TODO what about the rest of the $envValues from the imports and defs?
-      const CommandsBeforeEnv = _.merge({
-        iidy: {
-          stackArgs: stackArgsPass1,
-          stackName: argv.stackName || stackArgsPass1.StackName
-        }
-      }, argsdataPass1.$envValues);
-
-      // We want `iidy render` to show the results of that pre-processing:
-      argsdata.CommandsBefore = runCommandSet(
-        argsdata.CommandsBefore,
-        pathmod.dirname(argsfile),
-        CommandsBeforeEnv);
-    } else {
-      // Not on an iidy command that require CommandsBefore to be processed
-      // TODO ... do something more sensible here, such as escaping the commands
-      delete argsdata.CommandsBefore;
-    }
-  }
-  const stackArgsPass2 = await transform(argsdata, argsfile) as StackArgs;
-
-  showArgsfileWarnings(stackArgsPass2, argsfile);
-
-  const stackArgsPass3 = recursivelyMapValues(stackArgsPass2, (value: any) => {
-    if (typeof value === 'string') {
-      // $0string is an encoding added in preprocess/index.ts:visitString
-      return value.replace(/\$0string (0\d+)/g, '$1');
-    } else {
-      return value;
-    }
-  });
-  logger.debug('argsdata -> stackArgs', argsdata, '\n', stackArgsPass3);
-  return stackArgsPass3;
-};
-
-function recursivelyMapValues<T extends object>(o: T, f: (a: any) => any): T {
-  return _.mapValues(o, function(a: any) {
-    if(_.isArray(a)) {
-      return _.map(a, f);
-    } else if(_.isObject(a)) {
-      return recursivelyMapValues(a, f);
-    } else {
-      return f(a);
-    }
-  }) as T;
-}
-
-async function stackArgsToCreateStackInput(stackArgs: StackArgs, argsFilePath: string, environment: string, stackName?: string)
-  : Promise<aws.CloudFormation.CreateStackInput> {
-  let templateLocation;
-
-  if (stackArgs.ApprovedTemplateLocation) {
-    const approvedLocation = await approvedTemplateVersionLocation(
-      stackArgs.ApprovedTemplateLocation,
-      stackArgs.Template,
-      argsFilePath,
-      environment
-    );
-    templateLocation = `https://s3.amazonaws.com/${approvedLocation.Bucket}/${approvedLocation.Key}`;
-  } else {
-    templateLocation = stackArgs.Template;
-  }
-  // Template is optional for updates and update changesets
-  const {TemplateBody, TemplateURL} = await loadCFNTemplate(templateLocation, argsFilePath, environment);
-  const {StackPolicyBody, StackPolicyURL} = await loadCFNStackPolicy(stackArgs.StackPolicy, argsFilePath);
-
-  // TODO: DisableRollback
-  // specify either DisableRollback or OnFailure, but not both
-  const OnFailure = def('ROLLBACK', stackArgs.OnFailure)
-
-  if (stackArgs.ApprovedTemplateLocation) {
-    logger.debug(`ApprovedTemplateLocation: ${stackArgs.ApprovedTemplateLocation}`);
-    logger.debug(`Original Template: ${stackArgs.Template}`);
-    logger.debug(`TemplateURL with ApprovedTemplateLocation: ${TemplateURL}`);
-  } else {
-    logger.debug(`TemplateURL: ${TemplateURL}`);
-  }
-  if (stackArgs.RoleARN) {
-    logger.warn('RoleARN in stack-args.yaml is deprecated. Use ServiceRoleARN');
-  }
-
-  return {
-    StackName: def(stackArgs.StackName, stackName),
-    Capabilities: stackArgs.Capabilities,
-    NotificationARNs: stackArgs.NotificationARNs,
-    RoleARN: stackArgs.ServiceRoleARN || stackArgs.RoleARN,
-    OnFailure,
-    TimeoutInMinutes: stackArgs.TimeoutInMinutes,
-    ResourceTypes: stackArgs.ResourceTypes,
-    Parameters: objectToCFNParams(def({}, stackArgs.Parameters)),
-    Tags: objectToCFNTags(def({}, stackArgs.Tags)),
-    TemplateBody,
-    TemplateURL,
-    StackPolicyBody,
-    StackPolicyURL,
-    ClientRequestToken: stackArgs.ClientRequestToken
-  };
-}
-
-async function stackArgsToUpdateStackInput(
-  stackArgs: StackArgs,
-  argsFilePath: string,
-  environment: string,
-  stackName?: string)
-  : Promise<aws.CloudFormation.UpdateStackInput> {
-  const input0 = await stackArgsToCreateStackInput(stackArgs, argsFilePath, environment, stackName);
-  delete input0.TimeoutInMinutes;
-  delete input0.OnFailure;
-  const input = input0 as aws.CloudFormation.UpdateStackInput;
-  input.UsePreviousTemplate = stackArgs.UsePreviousTemplate;
-  return input0;
-}
-
-async function stackArgsToCreateChangeSetInput(
-  changeSetName: string,
-  stackArgs: StackArgs,
-  argsFilePath: string,
-  environment: string,
-  stackName?: string)
-  : Promise<aws.CloudFormation.CreateChangeSetInput> {
-  // TODO: ResourceTypes optionally locked down for changeset
-  const input0 = await stackArgsToCreateStackInput(stackArgs, argsFilePath, environment, stackName);
-  delete input0.TimeoutInMinutes;
-  delete input0.OnFailure;
-  delete input0.StackPolicyBody;
-  delete input0.StackPolicyURL;
-  const ClientToken = input0.ClientRequestToken; // damn CFN has inconsistent naming here
-  delete input0.ClientRequestToken;
-  const input = input0 as aws.CloudFormation.CreateChangeSetInput;
-  input.ChangeSetName = changeSetName;
-  input.ClientToken = ClientToken;
-  return input;
-}
-
-export function showFinalComandSummary(wasSuccessful: boolean): number {
-  if (wasSuccessful) {
-    console.log(formatSectionHeading(sprintf(`%-${COLUMN2_START}s`, 'Command Summary:')),
-      cli.black(cli.bgGreenBright('Success')),
-      '👍')
-    return SUCCESS;
-  } else {
-    console.log(
-      formatSectionHeading(sprintf(`%-${COLUMN2_START}s`, 'Command Summary:')),
-      cli.bgRedBright('Failure'),
-      ' (╯°□°）╯︵ ┻━┻ ',
-      'Fix and try again.'
-    )
-    return FAILURE;
-  }
-}
-
-async function isHttpTemplateAccessible(location?: string) {
-  if (location) {
-    try {
-      await request.get(location);
-      return true;
-    } catch (e) {
-      return false;
-    }
-  } else {
-    return false;
-  }
-}
+import {CfnOperation, StackArgs} from './types';
+import {showPendingChangesets} from './showPendingChangesets';
+import {EstimateStackCost} from './estimateStackCost';
 
 export async function doesStackExist(StackName: string): Promise<boolean> {
   const cfn = new aws.CloudFormation();
   return await cfn.describeStacks({StackName}).promise().thenReturn(true).catchReturn(false);
-}
-
-abstract class AbstractCloudFormationStackCommand {
-  public region: AWSRegion
-
-  readonly profile?: string;
-  readonly assumeRoleArn?: string;
-  readonly stackName: string;
-  readonly argsfile: string;
-  readonly environment: string;
-
-  protected cfnOperation: CfnOperation;
-  protected startTime: Date;
-  protected cfn: aws.CloudFormation;
-  protected expectedFinalStackStatus: string[];
-  protected showTimesInSummary: boolean = true;
-  protected showPreviousEvents: boolean = true;
-  protected previousStackEventsPromise: Promise<aws.CloudFormation.StackEvents>;
-  protected watchStackEvents: boolean = true;
-
-  constructor(readonly argv: GenericCLIArguments, readonly stackArgs: StackArgs) {
-    // region, profile, and assumeRoleArn are the only used for cli output here
-    // configureAWS is called by loadStackArgs prior to this constructor
-    // TODO We should cleanup / dry-up the resolution rules for these.
-    this.region = def(getCurrentAWSRegion(), this.argv.region || this.stackArgs.Region);
-    this.profile = (            // the resolved profile
-      this.argv.profile || this.stackArgs.Profile
-      || process.env.AWS_PROFILE || process.env.AWS_DEFAULT_PROFILE); // tslint:disable-line
-    this.assumeRoleArn = this.argv.assumeRoleArn || this.stackArgs.AssumeRoleARN;// tslint:disable-line
-
-    this.stackName = this.argv.stackName || this.stackArgs.StackName;// tslint:disable-line
-    this.argsfile = argv.argsfile;
-    this.environment = argv.environment;
-  }
-
-  async _setup() {
-    this.cfn = new aws.CloudFormation()
-    if (this.showPreviousEvents) {
-      this.previousStackEventsPromise = getAllStackEvents(this.stackName);
-    }
-  }
-
-  async _updateStackTerminationPolicy() {
-    if (_.isBoolean(this.stackArgs.EnableTerminationProtection)) {
-      const cfn = new aws.CloudFormation();
-      return cfn.updateTerminationProtection({
-        StackName: this.stackName,
-        EnableTerminationProtection: this.stackArgs.EnableTerminationProtection
-      }).promise();
-    }
-  }
-
-  async _showCommandSummary() {
-    const sts = new aws.STS();
-    const iamIdentPromise = sts.getCallerIdentity().promise();
-    const roleARN = this.stackArgs.ServiceRoleARN || this.stackArgs.RoleARN;
-
-    console.log(); // blank line
-    console.log(formatSectionHeading('Command Metadata:'))
-    printSectionEntry('CFN Operation:', cli.magenta(this.cfnOperation));
-    printSectionEntry('iidy Environment:', cli.magenta(this.environment));
-    printSectionEntry('Region:', cli.magenta(this.region));
-    if (!_.isEmpty(this.profile)) {
-      printSectionEntry('Profile:', cli.magenta(this.profile));
-    }
-
-    printSectionEntry(
-      'CLI Arguments:',
-      cli.blackBright(prettyFormatSmallMap(_.pick(this.argv, ['region', 'profile', 'argsfile']) as any)));
-
-    printSectionEntry('IAM Service Role:', cli.blackBright(def('None', roleARN)));
-
-    const iamIdent = await iamIdentPromise;
-    printSectionEntry('Current IAM Principal:', cli.blackBright(iamIdent.Arn));
-    printSectionEntry('iidy Version:', cli.blackBright(require('../../package.json').version));
-
-    console.log();
-  }
-
-  async run(): Promise<number> {
-    await this._setup();
-    await this._showCommandSummary();
-    this.startTime = await getReliableStartTime();
-    return this._run();
-  }
-
-  async _watchAndSummarize(stackId: string): Promise<number> {
-    // Show user all the meta data and stack properties
-    // TODO previous related stack, long-lived dependency stack, etc.
-
-    // we use StackId below rather than StackName to be resilient to deletions
-    const stackPromise = getStackDescription(stackId);
-    await summarizeStackDefinition(stackId, this.region, this.showTimesInSummary, stackPromise);
-
-    if (this.showPreviousEvents) {
-      console.log();
-      console.log(formatSectionHeading('Previous Stack Events (max 10):'))
-      await showStackEvents(stackId, 10, this.previousStackEventsPromise);
-    }
-
-    console.log();
-    if (this.watchStackEvents) {
-      await watchStack(stackId, this.startTime);
-    }
-
-    console.log();
-    const stack = await summarizeStackContents(stackId);
-
-    return showFinalComandSummary(_.includes(this.expectedFinalStackStatus, stack.StackStatus));
-  }
-
-  async _run(): Promise<number> {
-    throw new Error('Not implemented');
-  }
-
-  async _runCreate() {
-    if (_.isEmpty(this.stackArgs.Template)) {
-      throw new Error('For create-stack you must provide at Template: parameter in your argsfile')
-    };
-    try {
-      const createStackInput = await stackArgsToCreateStackInput(this.stackArgs, this.argsfile, this.environment, this.stackName);
-      if (await this._requiresTemplateApproval(createStackInput.TemplateURL)) {
-        return this._exitWithTemplateApprovalFailure();
-      }
-      const createStackOutput = await this.cfn.createStack(createStackInput).promise();
-      await this._updateStackTerminationPolicy();
-      return this._watchAndSummarize(createStackOutput.StackId as string);
-    } catch (e) {
-      if (e.message === 'CreateStack cannot be used with templates containing Transforms.') {
-        logger.error(
-          `Your stack template contains an AWS:: Transform so you need to use 'iidy create-or-update ${cli.red('--changeset')}'`)
-        return INTERRUPT;
-      } else {
-        throw e;
-      }
-    }
-  }
-
-  async _runUpdate() {
-    try {
-      let updateStackInput = await stackArgsToUpdateStackInput(this.stackArgs, this.argsfile, this.environment, this.stackName);
-      if (await this._requiresTemplateApproval(updateStackInput.TemplateURL)) {
-        return this._exitWithTemplateApprovalFailure();
-      }
-      if (this.argv.stackPolicyDuringUpdate) {
-        const {
-          StackPolicyBody: StackPolicyDuringUpdateBody,
-          StackPolicyURL: StackPolicyDuringUpdateURL
-        } = await loadCFNStackPolicy(this.argv.stackPolicyDuringUpdate as string, pathmod.join(process.cwd(), 'dummyfile'));
-        updateStackInput = _.merge({StackPolicyDuringUpdateBody, StackPolicyDuringUpdateURL}, updateStackInput);
-      }
-      await this._updateStackTerminationPolicy();
-      // TODO consider conditionally calling setStackPolicy if the policy has changed
-      const updateStackOutput = await this.cfn.updateStack(updateStackInput).promise();
-      return this._watchAndSummarize(updateStackOutput.StackId as string);
-    } catch (e) {
-      if (e.message === 'No updates are to be performed.') {
-        logger.info('No changes detected so no stack update needed.');
-        return SUCCESS;
-      } else if (e.message === 'UpdateStack cannot be used with templates containing Transforms.') {
-        const command = this.normalizeIidyCLICommand(`update-stack ${cli.red('--changeset')}'`);
-        logger.error(`Your stack template contains an AWS:: Transform so you need to use '${command}'`);
-        return INTERRUPT;
-      } else {
-        throw e;
-      }
-    }
-  }
-
-  async _requiresTemplateApproval(TemplateURL?: string): Promise<boolean> {
-    return !!(this.stackArgs.ApprovedTemplateLocation && ! await isHttpTemplateAccessible(TemplateURL));
-  }
-
-  _exitWithTemplateApprovalFailure(): number {
-    logger.error('Template version has not been approved or the current IAM principal does not have permission to access it. Run:');
-    logger.error(`  iidy template-approval request ${this.argsfile}`);
-    logger.error('to begin the approval process.');
-    return FAILURE;
-  }
-
-  normalizeIidyCLICommand(command: string): string {
-    let cliArgs = `--region ${this.region}`;
-    if (this.profile) {
-      cliArgs += ` --profile ${this.profile}`;
-    }
-    return `iidy ${cliArgs} ${command}`;
-  }
-
 }
 
 class CreateStack extends AbstractCloudFormationStackCommand {
@@ -688,53 +47,7 @@ class UpdateStack extends AbstractCloudFormationStackCommand {
   }
 }
 
-function summarizeChangeSet(changeSet: aws.CloudFormation.DescribeChangeSetOutput) {
-  for (const change of changeSet.Changes || []) {
-    if (change.ResourceChange) {
-      const resourceChange = change.ResourceChange;
-      const sprintfTemplate = '  %-17s %-30s %s';
-      switch (resourceChange.Action) {
-        case 'Add':
-          console.log(
-            sprintf(sprintfTemplate,
-              cli.green('Add'),
-              resourceChange.LogicalResourceId,
-              cli.blackBright(resourceChange.ResourceType)));
-          break;
-        case 'Remove':
-          console.log(
-            sprintf(sprintfTemplate,
-              cli.red('Remove'),
-              resourceChange.LogicalResourceId,
-              cli.blackBright(resourceChange.ResourceType + ' ' + resourceChange.PhysicalResourceId)));
-          break;
-        case 'Modify':
-          if (_.includes(['True', 'Conditional'], resourceChange.Replacement)) {
-            console.log(
-              sprintf(sprintfTemplate,
-                cli.red('Replace' + (resourceChange.Replacement === 'Conditional' ? '?' : '')),
-                resourceChange.LogicalResourceId,
-                cli.blackBright(resourceChange.ResourceType + ' ' + resourceChange.PhysicalResourceId)
-              ));
-          } else {
-            console.log(
-              sprintf(sprintfTemplate,
-                cli.yellow('Modify'),
-                resourceChange.LogicalResourceId,
-                cli.yellow(resourceChange.Scope || ''),
-                cli.blackBright(resourceChange.ResourceType + ' ' + resourceChange.PhysicalResourceId)
-              ));
-          }
-          if (resourceChange.Details) {
-            console.log(cli.blackBright(yaml.dump(resourceChange.Details)));
-          }
-          break;
-      }
-    }
-  }
-}
-
-class CreateChangeSet extends AbstractCloudFormationStackCommand {
+export class CreateChangeSet extends AbstractCloudFormationStackCommand {
   public changeSetName: string;
   public hasChanges: undefined | boolean;
 
@@ -830,18 +143,6 @@ class ExecuteChangeSet extends AbstractCloudFormationStackCommand {
   }
 }
 
-class EstimateStackCost extends AbstractCloudFormationStackCommand {
-  cfnOperation: CfnOperation = 'ESTIMATE_COST'
-
-  async _run() {
-    const {TemplateBody, TemplateURL, Parameters} =
-      await stackArgsToCreateStackInput(this.stackArgs, this.argsfile, this.environment, this.stackName)
-    const estimateResp = await this.cfn.estimateTemplateCost({TemplateBody, TemplateURL, Parameters}).promise();
-    console.log('Stack cost estimator: ', estimateResp.Url);
-    return SUCCESS;
-  }
-}
-
 const wrapCommandCtor =
   (Ctor: new (argv: GenericCLIArguments, stackArgs: StackArgs) => AbstractCloudFormationStackCommand) =>
     async function(argv: GenericCLIArguments): Promise<number> {
@@ -851,14 +152,6 @@ const wrapCommandCtor =
 export const createStackMain = wrapCommandCtor(CreateStack);
 export const executeChangesetMain = wrapCommandCtor(ExecuteChangeSet);
 export const estimateCost = wrapCommandCtor(EstimateStackCost);
-
-export function parseTemplateBody(TemplateBody: string): object {
-  if (TemplateBody.match(/^ *\{/) !== null) {
-    return JSON.parse(TemplateBody);
-  } else {
-    return yaml.loadString(TemplateBody, '');
-  }
-}
 
 export async function createOrUpdateStackMain(argv: GenericCLIArguments): Promise<number> {
   const stackArgs = await loadStackArgs(argv);
@@ -931,55 +224,3 @@ export async function updateStackMain(argv: GenericCLIArguments, stackArgs?: Sta
     return new UpdateStack(argv, stackArgs).run();
   }
 };
-
-
-export async function createChangesetMain(argv: GenericCLIArguments): Promise<number> {
-  const changesetRunner = new CreateChangeSet(argv, await loadStackArgs(argv));
-  const changesetExitCode = await changesetRunner.run();
-  if (argv.watch && changesetExitCode === 0) {
-    console.log();
-    await watchStack(changesetRunner.stackName, new Date(), DEFAULT_EVENT_POLL_INTERVAL, argv.watchInactivityTimeout);
-    console.log();
-    await summarizeStackContents(changesetRunner.stackName);
-    return SUCCESS;
-  } else {
-    return changesetExitCode;
-  }
-};
-
-
-export async function getStackNameFromArgsAndConfigureAWS(argv: GenericCLIArguments): Promise<string> {
-  if (/.*\.(yaml|yml)$/.test(argv.stackname) && fs.existsSync(argv.stackname)) {
-    const stackArgs = await _loadStackArgs(argv.stackname, argv);
-    return stackArgs.StackName;
-  } else {
-    await configureAWS(argv);
-    return argv.stackname;
-  }
-}
-
-export async function approvedTemplateVersionLocation(
-  approvedTemplateLocation: string,
-  templatePath: string,
-  baseLocation: string,
-  environment: string): Promise<{Bucket: string, Key: string}> {
-  // const templatePath = path.resolve(path.dirname(location), templatePath);
-  // const cfnTemplate = await fs.readFileSync(path.resolve(path.dirname(location), templatePath));
-  const cfnTemplate = await loadCFNTemplate(templatePath, baseLocation, environment, {omitMetadata: true});
-
-  if (cfnTemplate && cfnTemplate.TemplateBody) {
-    const s3Url = url.parse(approvedTemplateLocation);
-    const s3Path = s3Url.path ? s3Url.path : "";
-    const s3Bucket = s3Url.hostname ? s3Url.hostname : "";
-
-    const fileName = new Md5().appendStr(cfnTemplate.TemplateBody.toString()).end().toString()
-    const fullFileName = `${fileName}${pathmod.extname(templatePath)}`
-
-    return {
-      Bucket: s3Bucket,
-      Key: pathmod.join(s3Path.substring(1), fullFileName)
-    };
-  } else {
-    throw new Error('Unable to determine versioned template location');
-  }
-}
