+++ src/cfn/summarizeStackContents.ts
@@ -9,12 +9,12 @@ import {
   formatLogicalId,
   formatSectionHeading,
   formatStackExportName,
-  formatStackOutputName
+  formatStackOutputName,
+  calcPadding
 } from './formatting';
 import {getAllStackExportsWithImports} from './getAllStackExportsWithImports';
 import {getStackDescription} from './getStackDescription';
-import {showPendingChangesets} from './index';
-
+import {showPendingChangesets} from "./showPendingChangesets";
 
 export async function summarizeStackContents(StackId: string, stackPromise?: Promise<aws.CloudFormation.Stack>): Promise<aws.CloudFormation.Stack> {
   // TODO handle this part for when OnFailure=DELETE and stack is gone ...
@@ -25,18 +25,20 @@ export async function summarizeStackContents(StackId: string, stackPromise?: Pro
   const changeSetsPromise = cfn.listChangeSets({StackName: StackId}).promise();
   const stack = await (stackPromise || getStackDescription(StackId));
   const resources = def([], (await resourcesPromise).StackResources);
-  const MAX_PADDING = 60;
   if (resources.length > 0) {
     console.log(formatSectionHeading('Stack Resources:'));
-    const idPadding = Math.min(_.max(_.map(resources, r => r.LogicalResourceId.length)) as number, MAX_PADDING);
-    const resourceTypePadding = Math.min(_.max(_.map(resources, r => r.ResourceType.length)) as number, MAX_PADDING);
+    const idPadding = calcPadding(resources, r => r.LogicalResourceId);
+    const resourceTypePadding = calcPadding(resources, r => r.ResourceType);
     for (const resource of resources) {
-      console.log(formatLogicalId(sprintf(` %-${idPadding}s`, resource.LogicalResourceId)), cli.blackBright(sprintf(`%-${resourceTypePadding}s`, resource.ResourceType)), cli.blackBright(resource.PhysicalResourceId));
+      console.log(
+        formatLogicalId(sprintf(` %-${idPadding}s`, resource.LogicalResourceId)),
+        cli.blackBright(sprintf(`%-${resourceTypePadding}s`, resource.ResourceType)),
+        cli.blackBright(resource.PhysicalResourceId));
     }
   }
   console.log();
   process.stdout.write(formatSectionHeading('Stack Outputs:'));
-  const outputKeyPadding = Math.min(_.max(_.map(stack.Outputs, (output) => (output.OutputKey as string).length)) as number, MAX_PADDING);
+  const outputKeyPadding = calcPadding(stack.Outputs || [], o => o.OutputKey!);
   if (!_.isUndefined(stack.Outputs) && stack.Outputs.length > 0) {
     process.stdout.write('\n');
     for (const {OutputKey, OutputValue} of stack.Outputs) {
@@ -50,7 +52,7 @@ export async function summarizeStackContents(StackId: string, stackPromise?: Pro
   if (exports.length > 0) {
     console.log();
     console.log(formatSectionHeading('Stack Exports:'));
-    const exportNamePadding = Math.min(_.max(_.map(exports, (ex) => (ex.Name as string).length)) as number, MAX_PADDING);
+    const exportNamePadding = calcPadding(exports, ex => ex.Name!);
     for (const ex of exports) {
       console.log(formatStackExportName(sprintf(` %-${exportNamePadding}s`, ex.Name)), cli.blackBright(ex.Value));
       // TODO handle NextToken, which might happen on large sets of exports
@@ -61,7 +63,9 @@ export async function summarizeStackContents(StackId: string, stackPromise?: Pro
     }
   }
   console.log();
-  console.log(formatSectionHeading(sprintf(`%-${COLUMN2_START}s`, 'Current Stack Status:')), colorizeResourceStatus(stack.StackStatus), def('', stack.StackStatusReason));
+  console.log(formatSectionHeading(sprintf(`%-${COLUMN2_START}s`, 'Current Stack Status:')),
+              colorizeResourceStatus(stack.StackStatus),
+              def('', stack.StackStatusReason));
   await showPendingChangesets(StackId, changeSetsPromise);
   return stack;
 }
