+++ src/tests/test-cfn.ts
@@ -1,6 +1,6 @@
 require('./support'); // for side-effect
 import {expect} from 'chai';
-import {loadStackArgs} from '../cfn';
+import {loadStackArgs} from "../cfn/loadStackArgs";
 import * as aws from 'aws-sdk'
 
 describe('cfn', () => {
